location-matcher/requirements.txt
aiohttp==3.10.5
attr==0.3.2
brotli==1.1.0
brotlicffi==1.1.0.0
ConfigParser==7.1.0
contextlib2==21.6.0
cryptography==43.0.1
docutils==0.21.2
h2==4.1.0
HTMLParser==0.0.2
importlib_metadata==8.5.0
ipython==8.12.3
ipywidgets==8.1.5
Jinja2==3.1.4
jnius==1.1.0
js==1.0
keyring==25.4.1
lockfile==0.12.2
mock==5.1.0
Pillow==10.4.0
pyodide==0.0.2
pyOpenSSL==24.2.1
pytz==2024.2
railroad==0.5.0
requests_oauthlib==2.0.0
simplejson==3.19.3
Sphinx==8.0.2
thread==2.0.5
tornado==6.4.1
truststore==0.9.2
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
yarl==1.11.1
zstandard==0.23.0
